CRUD

+++ PMSUMail/app/src/main/java/com/example/pmsumail/FolderActivity.java
@@ -1,6 +1,5 @@
 package com.example.pmsumail;
 
-import android.content.Context;
 import android.content.Intent;
 import android.content.SharedPreferences;
 import android.os.Bundle;
@@ -9,21 +8,16 @@ import android.support.v7.widget.Toolbar;
 import android.view.Menu;
 import android.view.MenuInflater;
 import android.view.MenuItem;
-import android.view.View;
-import android.widget.AdapterView;
-import android.widget.ListView;
 import android.widget.TextView;
 import android.widget.Toast;
 
 import com.example.pmsumail.adapters.MessagesListAdapter;
-import com.example.pmsumail.model.Message;
 import com.example.pmsumail.model.Folder;
+import com.example.pmsumail.model.Message;
 import com.example.pmsumail.service.AccountService;
 import com.example.pmsumail.service.FolderService;
 import com.example.pmsumail.service.ServiceUtils;
 
-import java.io.ByteArrayOutputStream;
-import java.io.FileOutputStream;
 import java.util.ArrayList;
 
 import retrofit2.Call;
@@ -38,8 +32,10 @@ public class FolderActivity extends AppCompatActivity {
     private SharedPreferences sharedPreferences;
     String accountPrefe;
 
+    private TextView folderName;
 
-    private Folder folder = new Folder();
+
+    private Folder folder;
 
     private ArrayList<Folder> folders = new ArrayList<Folder>();
     private ArrayList<Message> messages = new ArrayList<Message>();
@@ -52,15 +48,24 @@ public class FolderActivity extends AppCompatActivity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_folder);
-
-
         Toolbar toolbar = findViewById(R.id.folder_toolbar);
         setSupportActionBar(toolbar);
 
-/*
-        TextView FolderName_view = findViewById(R.id.folder_title_view);
+        Bundle extras = getIntent().getExtras();
+        if (extras != null) {
+            folder = extras.getParcelable("folder");
+        }
+        folderService = ServiceUtils.folderService;
+        initView();
+
+/*        TextView FolderName_view = findViewById(R.id.folder_title_view);
 
         FolderName_view.setText("Folder name: " + folder.getName());*/
+    }
+
+    private void initView() {
+        folderName = findViewById(R.id.folder_name);
+        folderName.setText(folder.getName());
     }
 
     @Override
@@ -71,27 +76,30 @@ public class FolderActivity extends AppCompatActivity {
     }
 
     @Override
-    public boolean onOptionsItemSelected(MenuItem item){
-        switch (item.getItemId()){
+    public boolean onOptionsItemSelected(MenuItem item) {
+        switch (item.getItemId()) {
             case R.id.edit_folder:
-                Toast.makeText(getBaseContext(), "Edit folder" , Toast.LENGTH_SHORT ).show();
+                Toast.makeText(getBaseContext(), "Edit folder", Toast.LENGTH_SHORT).show();
                 return true;
             case R.id.delete_folder:
                 deleteFolder();
-                Toast.makeText(this, "Deleted", Toast.LENGTH_SHORT).show();
-                Intent intent = new Intent(this, FoldersActivity.class);
-                startActivity(intent);
         }
         return super.onOptionsItemSelected(item);
     }
 
     public void deleteFolder() {
+        if (folder.getName().equalsIgnoreCase("inbox") ||
+                folder.getName().equalsIgnoreCase("drafts")) {
+            Toast.makeText(this, "You cant delete " + folder.getName() + " folder", Toast.LENGTH_SHORT).show();
+            return;
+        }
         Call<Folder> call = folderService.deleteFolder(folder.getId());
 
         call.enqueue(new Callback<Folder>() {
             @Override
             public void onResponse(Call<Folder> call, Response<Folder> response) {
-
+                Toast.makeText(FolderActivity.this, "Folder deleted", Toast.LENGTH_SHORT).show();
+                finish();
             }
 
             @Override
